massive additions, restructuring: simulation - iterate handles copying, converting mu<>vperp automatically, random generator to Particle, changed prevSim to function from derived class, added check to end early if all particles have escaped; added functions to clear/zero GPU memory for later use; satellite now manages its own data, function can remove zeroes before saving to disk, reconstruct when loading; added CSV and Particle Distribution generating classes, added API functions (need more), API for testing B, E elements; improved ownership model of EElems - host EField no longer tracks GPU instances; CUDA macro added to prevent misuse; other organization changes

+++ src/LogFile/LogFile.cpp
@@ -6,7 +6,7 @@ void LogFile::writeLogFileEntry(std::string logMessage)
 	std::string writeTxt;
 	std::stringstream ss;
 
-	ss << std::setprecision(10) << std::setw(11) << static_cast<float>(std::chrono::duration_cast<std::chrono::microseconds>(std::chrono::steady_clock::now() - timeStructs_m.at(0)->tp).count()) / 1000;
+	ss << std::setprecision(10) << std::setw(11) << static_cast<double>(std::chrono::duration_cast<std::chrono::microseconds>(std::chrono::steady_clock::now() - timeStructs_m.at(0)->tp).count()) / 1000;
 	std::string ssstr{ ss.str() };
 	while (ssstr.length() > 11)
 		ssstr.erase(std::ios::end);
